+++ src/server/devices/baseDeviceInterfaces/iLoadMeter.ts
@@ -1,0 +1,6 @@
+export interface iLoadMeter {
+  /**
+   * The current load power in Watts
+   */
+  readonly loadPower: number;
+}