+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Ubisys Actuator as a load-meter
+
 ## 2.9.16 (2023-06-16)
 
 * (theimo1221) Add missing Device capability